All functions and types

LaplaceRedux.AbstractDecomposition

LaplaceRedux.AbstractLaplace

LaplaceRedux.AbstractLaplace

LaplaceRedux.Curvature.CurvatureInterface

LaplaceRedux.Curvature.EmpiricalFisher

LaplaceRedux.Curvature.GGN

LaplaceRedux.EstimationParams

LaplaceRedux.EstimationParams

LaplaceRedux.FullHessian

LaplaceRedux.HessianStructure

LaplaceRedux.Kron

LaplaceRedux.KronDecomposed

LaplaceRedux.KronHessian

LaplaceRedux.Laplace

LaplaceRedux.Laplace

LaplaceRedux.LaplaceClassification

LaplaceRedux.LaplaceParams

LaplaceRedux.LaplaceRegression

LaplaceRedux.Posterior

LaplaceRedux.Posterior

LaplaceRedux.Prior

LaplaceRedux.Prior

Base.:*

Base.:*

Base.:+

Base.:+

Base.:+

Base.:==

Base.getindex

Base.getindex

Base.length

Flux.params

Flux.params

LaplaceRedux.Curvature.full_batched

LaplaceRedux.Curvature.full_batched

LaplaceRedux.Curvature.full_unbatched

LaplaceRedux.Curvature.full_unbatched

LaplaceRedux.Curvature.gradients

LaplaceRedux.Curvature.jacobians

LaplaceRedux.Curvature.jacobians_batched

LaplaceRedux.Curvature.jacobians_unbatched

LaplaceRedux.Data.toy_data_linear

LaplaceRedux.Data.toy_data_multi

LaplaceRedux.Data.toy_data_non_linear

LaplaceRedux.Data.toy_data_regression

LaplaceRedux._H_factor

LaplaceRedux._fit!

LaplaceRedux._fit!

LaplaceRedux._init_H

LaplaceRedux._weight_penalty

LaplaceRedux.approximate

LaplaceRedux.approximate

LaplaceRedux.clamp

LaplaceRedux.convert_subnetwork_indices

LaplaceRedux.decompose

LaplaceRedux.fit!

LaplaceRedux.fit!

LaplaceRedux.functional_variance

LaplaceRedux.functional_variance

LaplaceRedux.functional_variance

LaplaceRedux.get_loss_fun

LaplaceRedux.get_loss_type

LaplaceRedux.get_map_estimate

LaplaceRedux.get_prior_mean

LaplaceRedux.glm_predictive_distribution

LaplaceRedux.has_softmax_or_sigmoid_final_layer

LaplaceRedux.hessian_approximation

LaplaceRedux.instantiate_curvature!

LaplaceRedux.interleave

LaplaceRedux.inv_square_form

LaplaceRedux.log_det_posterior_precision

LaplaceRedux.log_det_prior_precision

LaplaceRedux.log_det_ratio

LaplaceRedux.log_likelihood

LaplaceRedux.log_marginal_likelihood

LaplaceRedux.logdetblock

LaplaceRedux.mm

LaplaceRedux.n_params

LaplaceRedux.n_params

LaplaceRedux.optimize_prior!

LaplaceRedux.outdim

LaplaceRedux.outdim

LaplaceRedux.posterior_covariance

LaplaceRedux.posterior_precision

LaplaceRedux.predict

LaplaceRedux.prior_precision

LaplaceRedux.probit

LaplaceRedux.validate_subnetwork_indices

LinearAlgebra.det

LinearAlgebra.logdet

MLJFlux.build

MLJFlux.build

MLJFlux.fitresult

MLJFlux.fitresult

MLJFlux.shape

MLJFlux.train

MLJFlux.train

MLJModelInterface.predict

MLJModelInterface.predict

LaplaceRedux.@zb

Exported functions

LaplaceRedux.Laplace — Type

Laplace

Concrete type for Laplace approximation. This type is a subtype of AbstractLaplace and is used to store all the necessary information for a Laplace approximation.

Fields

model::Flux.Chain: The model to be approximated.

likelihood::Symbol: The likelihood function to be used.

est_params::EstimationParams: The estimation parameters.

prior::Prior: The parameters defining prior distribution.

posterior::Posterior: The posterior distribution.

LaplaceRedux.Laplace — Method

Laplace(model::Any; likelihood::Symbol, kwargs...)

Outer constructor for Laplace approximation. This function constructs a Laplace object from a given model and likelihood function.

Arguments

model::Any: The model to be approximated (a Flux.Chain).

likelihood::Symbol: The likelihood function to be used. Possible values are :regression and :classification.

Keyword Arguments

See LaplaceParams for a description of the keyword arguments.

Returns

la::Laplace: The Laplace object.

Examples

using Flux, LaplaceRedux
nn = Chain(Dense(2,1))
la = Laplace(nn, likelihood=:regression)

LaplaceRedux.LaplaceClassification — Type

MLJBase.@mlj_model mutable struct LaplaceClassification <: MLJFlux.MLJFluxProbabilistic

A mutable struct representing a Laplace Classification model that extends the MLJFluxProbabilistic abstract type. It uses Laplace approximation to estimate the posterior distribution of the weights of a neural network.

The model is defined by the following default parameters for all MLJFlux models:

builder: a Flux model that constructs the neural network.

finaliser: a Flux model that processes the output of the neural network.

optimiser: a Flux optimiser.

loss: a loss function that takes the predicted output and the true output as arguments.

epochs: the number of epochs.

batch_size: the size of a batch.

lambda: the regularization strength.

alpha: the regularization mix (0 for all l2, 1 for all l1).

rng: a random number generator.

optimiser_changes_trigger_retraining: a boolean indicating whether changes in the optimiser trigger retraining.

acceleration: the computational resource to use.

The model also has the following parameters, which are specific to the Laplace approximation:

subset_of_weights: the subset of weights to use, either :all, :last_layer, or :subnetwork.

subnetwork_indices: the indices of the subnetworks.

hessian_structure: the structure of the Hessian matrix, either :full or :diagonal.

backend: the backend to use, either :GGN or :EmpiricalFisher.

σ: the standard deviation of the prior distribution.

μ₀: the mean of the prior distribution.

P₀: the covariance matrix of the prior distribution.

link_approx: the link approximation to use, either :probit or :plugin.

predict_proba: a boolean that select whether to predict probabilities or not.

ret_distr: a boolean that tells predict to either return distributions (true) objects from Distributions.jl or just the probabilities.

fit_prior_nsteps: the number of steps used to fit the priors.

LaplaceRedux.LaplaceRegression — Type

MLJBase.@mlj_model mutable struct LaplaceRegression <: MLJFlux.MLJFluxProbabilistic

A mutable struct representing a Laplace regression model that extends the MLJFlux.MLJFluxProbabilistic abstract type. It uses Laplace approximation to estimate the posterior distribution of the weights of a neural network.

The model is defined by the following default parameters for all MLJFlux models:

builder: a Flux model that constructs the neural network.

optimiser: a Flux optimiser.

loss: a loss function that takes the predicted output and the true output as arguments.

epochs: the number of epochs.

batch_size: the size of a batch.

lambda: the regularization strength.

alpha: the regularization mix (0 for all l2, 1 for all l1).

rng: a random number generator.

optimiser_changes_trigger_retraining: a boolean indicating whether changes in the optimiser trigger retraining.

acceleration: the computational resource to use.

The model also has the following parameters, which are specific to the Laplace approximation:

subset_of_weights: the subset of weights to use, either :all, :last_layer, or :subnetwork.

subnetwork_indices: the indices of the subnetworks.

hessian_structure: the structure of the Hessian matrix, either :full or :diagonal.

backend: the backend to use, either :GGN or :EmpiricalFisher.

σ: the standard deviation of the prior distribution.

μ₀: the mean of the prior distribution.

P₀: the covariance matrix of the prior distribution.

ret_distr: a boolean that tells predict to either return distributions (true) objects from Distributions.jl or just the probabilities.

fit_prior_nsteps: the number of steps used to fit the priors.

LaplaceRedux.fit! — Method

fit!(la::AbstractLaplace,data)

Fits the Laplace approximation for a data set. The function returns the number of observations (n_data) that were used to update the Laplace object. It does not return the updated Laplace object itself because the function modifies the input Laplace object in place (as denoted by the use of '!' in the function's name).

Examples

using Flux, LaplaceRedux
x, y = LaplaceRedux.Data.toy_data_linear()
data = zip(x,y)
nn = Chain(Dense(2,1))
la = Laplace(nn)
fit!(la, data)

LaplaceRedux.fit! — Method

Fit the Laplace approximation, with batched data.

LaplaceRedux.glm_predictive_distribution — Method

glm_predictive_distribution(la::AbstractLaplace, X::AbstractArray)

Computes the linearized GLM predictive.

Arguments

la::AbstractLaplace: A Laplace object.

X::AbstractArray: Input data.

Returns

normal_distr A normal distribution N(fμ,fvar) approximating the predictive distribution p(y|X) given the input data X.

fμ::AbstractArray: Mean of the predictive distribution. The output shape is column-major as in Flux.

fvar::AbstractArray: Variance of the predictive distribution. The output shape is column-major as in Flux.

Examples

```julia-repl using Flux, LaplaceRedux using LaplaceRedux.Data: toydatalinear x, y = toydatalinear() data = zip(x,y) nn = Chain(Dense(2,1)) la = Laplace(nn; likelihood=:classification) fit!(la, data) glmpredictivedistribution(la, hcat(x...))

LaplaceRedux.optimize_prior! — Method

optimize_prior!(
    la::AbstractLaplace; 
    n_steps::Int=100, lr::Real=1e-1,
    λinit::Union{Nothing,Real}=nothing,
    σinit::Union{Nothing,Real}=nothing
)

Optimize the prior precision post-hoc through Empirical Bayes (marginal log-likelihood maximization).

LaplaceRedux.posterior_covariance — Function

posterior_covariance(la::AbstractLaplace, P=la.P)

Computes the posterior covariance $∑$ as the inverse of the posterior precision: $\Sigma=P^{-1}$.

LaplaceRedux.posterior_precision — Function

posterior_precision(la::AbstractLaplace, H=la.posterior.H, P₀=la.prior.P₀)

Computes the posterior precision $P$ for a fitted Laplace Approximation as follows,

$P = \sum_{n=1}^N\nabla_{\theta}^2 \log p(\mathcal{D}_n|\theta)|_{\hat\theta} + \nabla_{\theta}^2 \log p(\theta)|_{\hat\theta}$

where $\sum_{n=1}^N\nabla_{\theta}^2\log p(\mathcal{D}_n|\theta)|_{\hat\theta}=H$ is the Hessian and $\nabla_{\theta}^2 \log p(\theta)|_{\hat\theta}=P_0$ is the prior precision and $\hat\theta$ is the MAP estimate.

LaplaceRedux.predict — Method

predict(la::AbstractLaplace, X::AbstractArray; link_approx=:probit, predict_proba::Bool=true)

Computes predictions from Bayesian neural network.

Arguments

la::AbstractLaplace: A Laplace object.

X::AbstractArray: Input data.

link_approx::Symbol=:probit: Link function approximation. Options are :probit and :plugin.

predict_proba::Bool=true: If true (default) apply a sigmoid or a softmax function to the output of the Flux model.

return_distr::Bool=false: if false (default), the function output either the direct output of the chain or pseudo-probabilities (if predict_proba= true). if true predict return a Bernoulli distribution in binary classification tasks and a categorical distribution in multiclassification tasks.

Returns

For classification tasks, LaplaceRedux provides different options: if retdistr is false: - fμ::AbstractArray: Mean of the predictive distribution if link function is set to :plugin, otherwise the probit approximation. The output shape is column-major as in Flux. if retdistr is true: - a Bernoulli distribution in binary classification tasks and a categorical distribution in multiclassification tasks. For regression tasks:

normal_distr::Distributions.Normal:the array of Normal distributions computed by glmpredictivedistribution.

Examples

using Flux, LaplaceRedux
using LaplaceRedux.Data: toy_data_linear
x, y = toy_data_linear()
data = zip(x,y)
nn = Chain(Dense(2,1))
la = Laplace(nn; likelihood=:classification)
fit!(la, data)
predict(la, hcat(x...))

MLJModelInterface.predict — Method

predict(model::LaplaceClassification, Xnew)

Predicts the class labels for new data using the LaplaceClassification model.

Arguments

model::LaplaceClassification: The trained LaplaceClassification model.

fitresult: the fitresult output produced by MLJFlux.fit!

Xnew: The new data to make predictions on.

Returns

An array of predicted class labels.

MLJModelInterface.predict — Method

predict(model::LaplaceRegression, Xnew)

Predict the output for new input data using a Laplace regression model.

Arguments

model::LaplaceRegression: The trained Laplace regression model.

the fitresult output produced by MLJFlux.fit!

Xnew: The new input data.

Returns

The predicted output for the new input data.

LaplaceRedux.Curvature.CurvatureInterface — Type

Base type for any curvature interface.

Internal functions

LaplaceRedux.AbstractDecomposition — Type

Abstract type of Hessian decompositions.

LaplaceRedux.AbstractLaplace — Type

Abstract base type for all Laplace approximations in this library. All subclasses implemented are parametric.

LaplaceRedux.AbstractLaplace — Method

(la::AbstractLaplace)(X::AbstractArray)

Calling a model with Laplace Approximation on an array of inputs is equivalent to explicitly calling the predict function.

LaplaceRedux.EstimationParams — Type

EstimationParams

Container for the parameters of a Laplace approximation.

Fields

subset_of_weights::Symbol: the subset of weights to consider. Possible values are :all, :last_layer, and :subnetwork.

subnetwork_indices::Union{Nothing,Vector{Vector{Int}}}: the indices of the subnetwork. Possible values are nothing or a vector of vectors of integers.

hessian_structure::HessianStructure: the structure of the Hessian. Possible values are :full and :kron or a concrete subtype of HessianStructure.

curvature::Union{Curvature.CurvatureInterface,Nothing}: the curvature interface. Possible values are nothing or a concrete subtype of CurvatureInterface.

LaplaceRedux.EstimationParams — Method

EstimationParams(params::LaplaceParams)

Extracts the estimation parameters from a LaplaceParams object.

LaplaceRedux.FullHessian — Type

Concrete type for full Hessian structure. This is the default structure.

LaplaceRedux.HessianStructure — Type

Abstract type for Hessian structure.

LaplaceRedux.Kron — Type

Kronecker-factored approximate curvature representation for a neural network model. Each element in kfacs represents two Kronecker factors (𝐆, 𝐀), such that the full block Hessian approximation would be approximated as 𝐀⊗𝐆.

LaplaceRedux.KronDecomposed — Type

KronDecomposed

Decomposed Kronecker-factored approximate curvature representation for a neural network model.

Decomposition is required to add the prior (diagonal matrix) to the posterior (KronDecomposed). It also has the benefits of reducing the costs for computation of inverses and log-determinants.

LaplaceRedux.KronHessian — Type

Concrete type for Kronecker-factored Hessian structure.

LaplaceRedux.LaplaceParams — Type

LaplaceParams

Container for the parameters of a Laplace approximation.

Fields

subset_of_weights::Symbol: the subset of weights to consider. Possible values are :all, :last_layer, and :subnetwork.

subnetwork_indices::Union{Nothing,Vector{Vector{Int}}}: the indices of the subnetwork. Possible values are nothing or a vector of vectors of integers.

hessian_structure::HessianStructure: the structure of the Hessian. Possible values are :full and :kron or a concrete subtype of HessianStructure.

backend::Symbol: the backend to use. Possible values are :GGN and :Fisher.

curvature::Union{Curvature.CurvatureInterface,Nothing}: the curvature interface. Possible values are nothing or a concrete subtype of CurvatureInterface.

σ::Real: the observation noise

μ₀::Real: the prior mean

λ::Real: the prior precision

P₀::Union{Nothing,AbstractMatrix,UniformScaling}: the prior precision matrix

LaplaceRedux.Posterior — Type

Posterior

Container for the results of a Laplace approximation.

Fields

μ::AbstractVector: the MAP estimate of the parameters

H::Union{AbstractArray,AbstractDecomposition,Nothing}: the Hessian matrix

P::Union{AbstractArray,AbstractDecomposition,Nothing}: the posterior precision matrix

Σ::Union{AbstractArray,Nothing}: the posterior covariance matrix

n_data::Union{Int,Nothing}: the number of data points

n_params::Union{Int,Nothing}: the number of parameters

n_out::Union{Int,Nothing}: the number of outputs

loss::Real: the loss value

LaplaceRedux.Posterior — Method

Posterior(model::Any, est_params::EstimationParams)

Outer constructor for Posterior object.

LaplaceRedux.Prior — Type

Prior

Container for the prior parameters of a Laplace approximation.

Fields

σ::Real: the observation noise

μ₀::Real: the prior mean

λ::Real: the prior precision

P₀::Union{Nothing,AbstractMatrix,UniformScaling}: the prior precision matrix

LaplaceRedux.Prior — Method

Prior(params::LaplaceParams)

Extracts the prior parameters from a LaplaceParams object.

Base.:* — Method

Multiply by a scalar by changing the eigenvalues. Distribute the scalar along the factors of a block.

Base.:* — Method

Kronecker-factored curvature scalar scaling.

Base.:+ — Method

Kronecker-factored curvature sum.

Base.:+ — Method

Shift the factors by a diagonal (assumed uniform scaling)

Base.:+ — Method

Shift the factors by a scalar across the diagonal.

Base.:== — Method

Kronecker-factored curvature equality.

Base.getindex — Method

Get Kronecker-factored block represenation.

Base.getindex — Method

Get i-th block of a a Kronecker-factored curvature.

Base.length — Method

Number of blocks in a Kronecker-factored curvature.

Flux.params — Method

Flux.params(model::Any, params::EstimationParams)

Extracts the parameters of a model based on the subset of weights specified in the EstimationParams object.

Flux.params — Method

Flux.params(la::Laplace)

Overloads the params function for a Laplace object.

LaplaceRedux._H_factor — Method

_H_factor(la::AbstractLaplace)

Returns the factor σ⁻², where σ is used in the zero-centered Gaussian prior p(θ) = N(θ;0,σ²I)

LaplaceRedux._fit! — Method

_fit!(la::Laplace, hessian_structure::FullHessian, data; batched::Bool=false, batchsize::Int, override::Bool=true)

Fit a Laplace approximation to the posterior distribution of a model using the full Hessian.

LaplaceRedux._fit! — Method

_fit!(la::Laplace, hessian_structure::KronHessian, data; batched::Bool=false, batchsize::Int, override::Bool=true)

Fit a Laplace approximation to the posterior distribution of a model using the Kronecker-factored Hessian.

LaplaceRedux._init_H — Method

_init_H(la::AbstractLaplace)

LaplaceRedux._weight_penalty — Method

_weight_penalty(la::AbstractLaplace)

The weight penalty term is a regularization term used to prevent overfitting. Weight regularization methods such as weight decay introduce a penalty to the loss function when training a neural network to encourage the network to use small weights. Smaller weights in a neural network can result in a model that is more stable and less likely to overfit the training dataset, in turn having better performance when making a prediction on new data.

LaplaceRedux.approximate — Method

approximate(curvature::CurvatureInterface, hessian_structure::FullHessian, d::Tuple; batched::Bool=false)

Compute the full approximation, for either a single input-output datapoint or a batch of such.

LaplaceRedux.approximate — Method

approximate(curvature::CurvatureInterface, hessian_structure::KronHessian, data; batched::Bool=false)

Compute the eigendecomposed Kronecker-factored approximate curvature as the Fisher information matrix.

Note, since the network predictive distribution is used in a weighted sum, and the number of backward passes is linear in the number of target classes, e.g. 100 for CIFAR-100.

LaplaceRedux.clamp — Method

Clamp eigenvalues in an eigendecomposition to be non-negative.

Since the Fisher information matrix is a positive-semidefinite by construction, the (near-zero) negative eigenvalues should be neglected.

LaplaceRedux.convert_subnetwork_indices — Method

convertsubnetworkindices(subnetwork_indices::AbstractArray)

Converts the subnetwork indices from the user given format [theta, row, column] to an Int i that corresponds to the index of that weight in the flattened array of weights.

LaplaceRedux.decompose — Method

decompose(K::Kron)

Eigendecompose Kronecker factors and turn into KronDecomposed.

LaplaceRedux.functional_variance — Method

functional_variance(la::AbstractLaplace, 𝐉::AbstractArray)

Compute the functional variance for the GLM predictive. Dispatches to the appropriate method based on the Hessian structure.

LaplaceRedux.functional_variance — Method

functional_variance(la::Laplace,𝐉)

Compute the linearized GLM predictive variance as 𝐉ₙΣ𝐉ₙ' where 𝐉=∇f(x;θ)|θ̂ is the Jacobian evaluated at the MAP estimate and Σ = P⁻¹.

LaplaceRedux.functional_variance — Method

functionalvariance(la::Laplace, hessianstructure::KronHessian, 𝐉::Matrix)

Compute functional variance for the GLM predictive: as the diagonal of the K×K predictive output covariance matrix 𝐉𝐏⁻¹𝐉ᵀ, where K is the number of outputs, 𝐏 is the posterior precision, and 𝐉 is the Jacobian of model output 𝐉=∇f(x;θ)|θ̂.

LaplaceRedux.get_loss_fun — Method

get_loss_fun(likelihood::Symbol)

Helper function to choose loss function based on specified model likelihood.

LaplaceRedux.get_loss_type — Method

get_loss_type(likelihood::Symbol)

Choose loss function type based on specified model likelihood.

LaplaceRedux.get_map_estimate — Method

get_map_estimate(model::Any, est_params::EstimationParams)

Helper function to extract the MAP estimate of the parameters for the model based on the subset of weights specified in the EstimationParams object.

LaplaceRedux.get_prior_mean — Method

get_prior_mean(la::Laplace)

Helper function to extract the prior mean of the parameters from a Laplace approximation.

LaplaceRedux.has_softmax_or_sigmoid_final_layer — Method

has_softmax_or_sigmoid_final_layer(model::Flux.Chain)

Check if the FLux model ends with a sigmoid or with a softmax layer

Input: - model: the Flux Chain object that represent the neural network. Return: - has_finaliser: true if the check is positive, false otherwise.

LaplaceRedux.hessian_approximation — Method

hessian_approximation(la::AbstractLaplace, d; batched::Bool=false)

Computes the local Hessian approximation at a single datapoint d.

LaplaceRedux.instantiate_curvature! — Method

instantiate_curvature!(params::EstimationParams, model::Any, likelihood::Symbol, backend::Symbol)

Instantiates the curvature interface for a Laplace approximation. The curvature interface is a concrete subtype of CurvatureInterface and is used to compute the Hessian matrix. The curvature interface is stored in the curvature field of the EstimationParams object.

LaplaceRedux.interleave — Method

Interleave elements of multiple iterables in order provided.

LaplaceRedux.inv_square_form — Method

function invsquareform(K::KronDecomposed, W::Matrix)

Special function to compute the inverse square form 𝐉𝐏⁻¹𝐉ᵀ (or 𝐖𝐊⁻¹𝐖ᵀ)

LaplaceRedux.log_det_posterior_precision — Method

log_det_posterior_precision(la::AbstractLaplace)

LaplaceRedux.log_det_prior_precision — Method

log_det_prior_precision(la::AbstractLaplace)

LaplaceRedux.log_det_ratio — Method

log_det_ratio(la::AbstractLaplace)

LaplaceRedux.log_likelihood — Method

log_likelihood(la::AbstractLaplace)

LaplaceRedux.log_marginal_likelihood — Method

log_marginal_likelihood(la::AbstractLaplace; P₀::Union{Nothing,UniformScaling}=nothing, σ::Union{Nothing, Real}=nothing)

LaplaceRedux.logdetblock — Method

logdetblock(block::Tuple{Eigen,Eigen}, delta::Number)

Log-determinant of a block in KronDecomposed, shifted by delta by on the diagonal.

LaplaceRedux.mm — Method

Matrix-multuply for the KronDecomposed Hessian approximation K and a 2-d matrix W, applying an exponent to K and transposing W before multiplication. Return (K^x)W^T, where x is the exponent.

LaplaceRedux.n_params — Method

n_params(model::Any, params::EstimationParams)

Helper function to determine the number of parameters of a Flux.Chain with Laplace approximation.

LaplaceRedux.n_params — Method

LaplaceRedux.n_params(la::Laplace)

Overloads the n_params function for a Laplace object.

LaplaceRedux.outdim — Method

outdim(model::Chain)

Helper function to determine the output dimension of a Flux.Chain, corresponding to the number of neurons on the last layer of the NN.

LaplaceRedux.outdim — Method

outdim(la::AbstractLaplace)

Helper function to determine the output dimension, corresponding to the number of neurons on the last layer of the NN, of a Flux.Chain with Laplace approximation.

LaplaceRedux.prior_precision — Method

prior_precision(la::Laplace)

Helper function to extract the prior precision matrix from a Laplace approximation.

LaplaceRedux.probit — Method

probit(fμ::AbstractArray, fvar::AbstractArray)

Compute the probit approximation of the predictive distribution.

LaplaceRedux.validate_subnetwork_indices — Method

validatesubnetworkindices( subnetwork_indices::Union{Nothing,Vector{Vector{Int}}}, params )

Determines whether subnetwork_indices is a valid input for specified parameters.

LinearAlgebra.det — Method

det(K::KronDecomposed)

Log-determinant of the KronDecomposed block-diagonal matrix, as the exponentiated log-determinant.

LinearAlgebra.logdet — Method

logdet(K::KronDecomposed)

Log-determinant of the KronDecomposed block-diagonal matrix, as the product of the determinants of the blocks

MLJFlux.build — Method

MLJFlux.build(model::LaplaceClassification, rng, shape)

Builds an MLJFlux model for Laplace classification compatible with the dimensions of the input and output layers specified by shape.

Arguments

model::LaplaceClassification: The Laplace classification model.

rng: A random number generator to ensure reproducibility.

shape: A tuple or array specifying the dimensions of the input and output layers.

Returns

The constructed MLJFlux model, compatible with the specified input and output dimensions.

MLJFlux.build — Method

MLJFlux.build(model::LaplaceRegression, rng, shape)

Builds an MLJFlux model for Laplace regression compatible with the dimensions of the input and output layers specified by shape.

Arguments

model::LaplaceRegression: The Laplace regression model.

rng: A random number generator to ensure reproducibility.

shape: A tuple or array specifying the dimensions of the input and output layers.

Returns

The constructed MLJFlux model, compatible with the specified input and output dimensions.

MLJFlux.fitresult — Method

MLJFlux.fitresult(model::LaplaceClassification, chain, y)

Computes the fit result for a Laplace classification model, returning the model chain and the number of unique classes in the target data.

Arguments

model::LaplaceClassification: The Laplace classification model to be evaluated.

chain: The trained model chain.

y: The target data, typically a vector of class labels.

Returns

Returns

A tuple containing:

The trained Flux chain.

a deepcopy of the laplace model.

MLJFlux.fitresult — Method

MLJFlux.fitresult(model::LaplaceRegression, chain, y)

Computes the fit result for a Laplace Regression model, returning the model chain and the number of output variables in the target data.

Arguments

model::LaplaceRegression: The Laplace Regression model to be evaluated.

chain: The trained model chain.

y: The target data, typically a vector of class labels.

Returns

A tuple containing:

The trained Flux chain.

a deepcopy of the laplace model.

MLJFlux.shape — Method

MLJFlux.shape(model::LaplaceRegression, X, y)

Compute the the number of features of the X input dataset and the number of variables to predict from the y output dataset.

Arguments

model::LaplaceRegression: The LaplaceRegression model to fit.

X: The input data for training.

y: The target labels for training one-hot encoded.

Returns

(input size, output size)

MLJFlux.train — Method

MLJFlux.train(model::LaplaceClassification, chain, regularized_optimiser, optimiser_state, epochs, verbosity, X, y)

Fit the LaplaceRegression model using Flux.jl.

Arguments

model::LaplaceClassification: The LaplaceClassification model.

regularized_optimiser: the regularized optimiser to apply to the loss function.

optimiser_state: thestate of the optimiser.

epochs: The number of epochs for training.

verbosity: The verbosity level for training.

X: The input data for training.

y: The target labels for training.

Returns (fitresult, cache, report )

where

la: the fitted Laplace model.

optimiser_state: the state of the optimiser.

history: the training loss history.

MLJFlux.train — Method

MLJFlux.train(model::LaplaceRegression, chain, regularized_optimiser, optimiser_state, epochs, verbosity, X, y)

Fit the LaplaceRegression model using Flux.jl.

Arguments

model::LaplaceRegression: The LaplaceRegression model.

regularized_optimiser: the regularized optimiser to apply to the loss function.

optimiser_state: thestate of the optimiser.

epochs: The number of epochs for training.

verbosity: The verbosity level for training.

X: The input data for training.

y: The target labels for training.

Returns (la, optimiser_state, history )

where

la: the fitted Laplace model.

optimiser_state: the state of the optimiser.

history: the training loss history.

LaplaceRedux.@zb — Macro

Macro for zero-based indexing. Example of usage: (@zb A[0]) = ...

LaplaceRedux.Curvature.EmpiricalFisher — Type

Constructor for curvature approximated by empirical Fisher.

LaplaceRedux.Curvature.GGN — Type

Constructor for curvature approximated by Generalized Gauss-Newton.

LaplaceRedux.Curvature.full_batched — Method

full_batched(curvature::EmpiricalFisher, d::Tuple)

Compute the full empirical Fisher for batch of inputs-outputs, with the batch dimension at the end.

LaplaceRedux.Curvature.full_batched — Method

full_batched(curvature::GGN, d::Tuple)

Compute the full GGN for batch of inputs-outputs, with the batch dimension at the end.

LaplaceRedux.Curvature.full_unbatched — Method

full_unbatched(curvature::EmpiricalFisher, d::Tuple)

Compute the full empirical Fisher for a single datapoint.

LaplaceRedux.Curvature.full_unbatched — Method

full_unbatched(curvature::GGN, d::Tuple)

Compute the full GGN for a singular input-ouput datapoint.

LaplaceRedux.Curvature.gradients — Method

gradients(curvature::CurvatureInterface, X::AbstractArray, y::Number)

Compute the gradients with respect to the loss function: ∇ℓ(f(x;θ),y) where f: ℝᴰ ↦ ℝᴷ.

LaplaceRedux.Curvature.jacobians — Method

jacobians(curvature::CurvatureInterface, X::AbstractArray; batched::Bool=false)

Computes the Jacobian ∇f(x;θ) where f: ℝᴰ ↦ ℝᴷ.

LaplaceRedux.Curvature.jacobians_batched — Method

jacobians_batched(curvature::CurvatureInterface, X::AbstractArray)

Compute Jacobians of the model output w.r.t. model parameters for points in X, with batching.

LaplaceRedux.Curvature.jacobians_unbatched — Method

jacobians_unbatched(curvature::CurvatureInterface, X::AbstractArray)

Compute the Jacobian of the model output w.r.t. model parameters for the point X, without batching. Here, the nn function is wrapped in an anonymous function using the () -> syntax, which allows it to be differentiated using automatic differentiation.

LaplaceRedux.Data.toy_data_linear — Function

toy_data_linear(N=100)

Examples

toy_data_linear()

LaplaceRedux.Data.toy_data_multi — Function

toy_data_multi(N=100)

Examples

toy_data_multi()

LaplaceRedux.Data.toy_data_non_linear — Function

toy_data_non_linear(N=100)

Examples

toy_data_non_linear()

LaplaceRedux.Data.toy_data_regression — Function

toy_data_regression(N=25, p=1; noise=0.3, fun::Function=f(x)=sin(2 * π * x))

A helper function to generate synthetic data for regression.