All functions and types

LaplaceRedux.BaseLaplace

LaplaceRedux.Curvature.CurvatureInterface

LaplaceRedux.Curvature.EmpiricalFisher

LaplaceRedux.Laplace

LaplaceRedux.Laplace

LaplaceRedux.Curvature.full

LaplaceRedux.Curvature.gradients

LaplaceRedux.Curvature.jacobians

LaplaceRedux._H_factor

LaplaceRedux._init_H

LaplaceRedux._weight_penalty

LaplaceRedux.fit!

LaplaceRedux.functional_variance

LaplaceRedux.get_loss_fun

LaplaceRedux.get_params

LaplaceRedux.glm_predictive_distribution

LaplaceRedux.hessian_approximation

LaplaceRedux.log_det_posterior_precision

LaplaceRedux.log_det_prior_precision

LaplaceRedux.log_det_ratio

LaplaceRedux.log_likelihood

LaplaceRedux.log_marginal_likelihood

LaplaceRedux.optimize_prior!

LaplaceRedux.outdim

LaplaceRedux.outdim

LaplaceRedux.posterior_covariance

LaplaceRedux.posterior_precision

LaplaceRedux.predict

Exported functions

LaplaceRedux.Laplace — Method

(la::Laplace)(X::AbstractArray; kwrgs...)

Calling a model with Laplace Approximation on an array of inputs is equivalent to explicitly calling the predict function.

LaplaceRedux.Laplace — Method

Laplace(model::Any; loss_fun::Union{Symbol, Function}, kwargs...)

Wrapper function to prepare Laplace approximation.

LaplaceRedux.fit! — Method

fit!(la::Laplace,data)

Fits the Laplace approximation for a data set.

Examples

using Flux, LaplaceRedux
x, y = LaplaceRedux.Data.toy_data_linear()
data = zip(x,y)
nn = Chain(Dense(2,1))
la = Laplace(nn)
fit!(la, data)

LaplaceRedux.optimize_prior! — Method

optimize_prior!(
    la::Laplace; 
    n_steps::Int=100, lr::Real=1e-1,
    λinit::Union{Nothing,Real}=nothing,
    σinit::Union{Nothing,Real}=nothing
)

Optimize the prior precision post-hoc through Empirical Bayes (marginal log-likelihood maximization).

LaplaceRedux.predict — Method

predict(la::Laplace, X::AbstractArray; link_approx=:probit)

Computes predictions from Bayesian neural network.

Examples

using Flux, LaplaceRedux
x, y = toy_data_linear()
data = zip(x,y)
nn = Chain(Dense(2,1))
la = Laplace(nn)
fit!(la, data)
predict(la, hcat(x...))

Internal functions

LaplaceRedux.BaseLaplace — Type

Abstract base type of Laplace Approximation.

LaplaceRedux._H_factor — Method

_H_factor(la::BaseLaplace)

LaplaceRedux._init_H — Method

_init_H(la::BaseLaplace)

LaplaceRedux._weight_penalty — Method

_weight_penalty(la::BaseLaplace)

LaplaceRedux.functional_variance — Method

functional_variance(la::Laplace,𝐉)

Compute the linearized GLM predictive variance as 𝐉ₙΣ𝐉ₙ' where 𝐉=∇f(x;θ)|θ̂ is the Jacobian evaluated at the MAP estimate and Σ = P⁻¹.

LaplaceRedux.get_loss_fun — Method

get_loss_fun(likelihood::Symbol)

Helper function to choose loss function based on specified model likelihood.

LaplaceRedux.get_params — Method

get_params(la::BaseLaplace)

Retrieves the desired (sub)set of model parameters and stores them in a list.

Examples

using Flux, LaplaceRedux
nn = Chain(Dense(2,1))
la = Laplace(nn)
LaplaceRedux.get_params(la)

LaplaceRedux.glm_predictive_distribution — Method

glm_predictive_distribution(la::Laplace, X::AbstractArray)

Computes the linearized GLM predictive.

LaplaceRedux.hessian_approximation — Method

hessian_approximation(la::Laplace, d)

Computes the local Hessian approximation at a single data d.

LaplaceRedux.log_det_posterior_precision — Method

log_det_posterior_precision(la::Laplace)

LaplaceRedux.log_det_prior_precision — Method

log_det_prior_precision(la::Laplace)

LaplaceRedux.log_det_ratio — Method

log_det_ratio(la::BaseLaplace)

LaplaceRedux.log_likelihood — Method

log_likelihood(la::BaseLaplace)

LaplaceRedux.log_marginal_likelihood — Method

log_marginal_likelihood(la::BaseLaplace; P₀::Union{Nothing,UniformScaling}=nothing, σ::Union{Nothing, Real}=nothing)

LaplaceRedux.outdim — Method

outdim(model::Chain)

Helper function to determine the output dimension of a Flux.Chain

LaplaceRedux.outdim — Method

outdim(la::BaseLaplace)

Helper function to determine the output dimension of a Flux.Chain with Laplace approximation.

LaplaceRedux.posterior_covariance — Function

posterior_covariance(la::BaseLaplace, P=la.P)

Computes the posterior covariance as the inverse of the posterior precision: $\Sigma=P^{-1}$.

LaplaceRedux.posterior_precision — Function

posterior_precision(la::BaseLaplace)

Computes the posterior precision $P$ for a fitted Laplace Approximation as follows,

$P = \sum_{n=1}^N\nabla_{\theta}^2\log p(\mathcal{D}_n|\theta)|_{\theta}_{MAP} + \nabla_{\theta}^2 \log p(\theta)|_{\theta}_{MAP}$

where $\sum_{n=1}^N\nabla_{\theta}^2\log p(\mathcal{D}_n|\theta)|_{\theta}_{MAP}=H$ and $\nabla_{\theta}^2 \log p(\theta)|_{\theta}_{MAP}=P_0$.

LaplaceRedux.Curvature.CurvatureInterface — Type

Basetype for any curvature interface.

LaplaceRedux.Curvature.EmpiricalFisher — Type

Constructor for Empirical Fisher.

LaplaceRedux.Curvature.full — Method

full(curvature::EmpiricalFisher, d::Union{Tuple,NamedTuple})

Compute the full empirical Fisher.

LaplaceRedux.Curvature.gradients — Method

gradients(curvature::CurvatureInterface, X::AbstractArray, y::Number)

Compute the gradients with respect to the loss function: ∇ℓ(f(x;θ),y) where f: ℝᴰ ↦ ℝᴷ.

LaplaceRedux.Curvature.jacobians — Method

jacobians(curvature::CurvatureInterface, X::AbstractArray)

Computes the Jacobian ∇f(x;θ) where f: ℝᴰ ↦ ℝᴷ.