All functions and types

LaplaceRedux.Curvature.CurvatureInterface

LaplaceRedux.Curvature.EmpiricalFisher

LaplaceRedux.Laplace

LaplaceRedux.Laplace

LaplaceRedux.Curvature.full

LaplaceRedux.Curvature.gradients

LaplaceRedux.Curvature.jacobians

LaplaceRedux.fit!

LaplaceRedux.get_loss_fun

LaplaceRedux.get_params

LaplaceRedux.glm_predictive_distribution

LaplaceRedux.hessian_approximation

LaplaceRedux.optimize_prior_precision

LaplaceRedux.outdim

LaplaceRedux.outdim

LaplaceRedux.predict

LaplaceRedux.predictive_variance

Exported functions

LaplaceRedux.Laplace — Method

(la::Laplace)(X::AbstractArray; kwrgs...)

Calling a model with Laplace Approximation on an array of inputs is equivalent to explicitly calling the predict function.

LaplaceRedux.Laplace — Method

Laplace(model::Any; loss_fun::Union{Symbol, Function}, kwargs...)

Wrapper function to prepare Laplace approximation.

LaplaceRedux.fit! — Method

fit!(la::Laplace,data)

Fits the Laplace approximation for a data set.

Examples

using Flux, LaplaceRedux
x, y = LaplaceRedux.Data.toy_data_linear()
data = zip(x,y)
nn = Chain(Dense(2,1))
la = Laplace(nn)
fit!(la, data)

LaplaceRedux.predict — Method

predict(la::Laplace, X::AbstractArray; link_approx=:probit)

Computes predictions from Bayesian neural network.

Examples

using Flux, LaplaceRedux
x, y = toy_data_linear()
data = zip(x,y)
nn = Chain(Dense(2,1))
la = Laplace(nn)
fit!(la, data)
predict(la, hcat(x...))

Internal functions

LaplaceRedux.get_loss_fun — Method

get_loss_fun(likelihood::Symbol)

Helper function to choose loss function based on specified model likelihood.

LaplaceRedux.get_params — Method

get_params(la::Laplace)

Retrieves the desired (sub)set of model parameters and stores them in a list.

Examples

using Flux, LaplaceRedux
nn = Chain(Dense(2,1))
la = Laplace(nn)
LaplaceRedux.get_params(la)

LaplaceRedux.glm_predictive_distribution — Method

glm_predictive_distribution(la::Laplace, X::AbstractArray)

Computes the linearized GLM predictive.

LaplaceRedux.hessian_approximation — Method

hessian_approximation(la::Laplace, d)

Computes the local Hessian approximation at a single data d.

LaplaceRedux.optimize_prior_precision — Method

optimize_prior_precision(la::Laplace; n_steps=100, lr=1e-1, init_prior_prec=1.)

Optimize the prior precision post-hoc through empirical Bayes (marginal log-likelihood maximization).

LaplaceRedux.outdim — Method

outdim(model::Chain)

Helper function to determine the output dimension of a Flux.Chain

LaplaceRedux.outdim — Method

outdim(la::Laplace)

Helper function to determine the output dimension of a Flux.Chain with Laplace approximation.

LaplaceRedux.predictive_variance — Method

predictive_variance(la::Laplace,𝐉)

Compute the linearized GLM predictive variance as 𝐉ₙΣ𝐉ₙ' where 𝐉=∇f(x;θ)|θ̂ is the Jacobian evaluated at the MAP estimate and Σ = H⁻¹.

LaplaceRedux.Curvature.CurvatureInterface — Type

Basetype for any curvature interface.

LaplaceRedux.Curvature.EmpiricalFisher — Type

Constructor for Empirical Fisher.

LaplaceRedux.Curvature.full — Method

full(curvature::EmpiricalFisher, d::Union{Tuple,NamedTuple})

Compute the full empirical Fisher.

LaplaceRedux.Curvature.gradients — Method

gradients(curvature::CurvatureInterface, X::AbstractArray, y::Number)

Compute the gradients with respect to the loss function: ∇ℓ(f(x;θ),y) where f: ℝᴰ ↦ ℝᴷ.

LaplaceRedux.Curvature.jacobians — Method

jacobians(curvature::CurvatureInterface, X::AbstractArray)

Computes the Jacobian ∇f(x;θ) where f: ℝᴰ ↦ ℝᴷ.